Update DateField component, add DateInput component
