+++ packages/design-system-docs/src/pages/components/DateField/datefield.example.html
@@ -15,7 +15,7 @@
           type="text"
           inputmode="numeric"
           pattern="[0-9]*"
-          id="month"
+          id="month-1"
           name="birthdate[month]"
           autocomplete="bday-month"
           aria-describedby="datefield_label_1"
@@ -30,7 +30,7 @@
           type="text"
           inputmode="numeric"
           pattern="[0-9]*"
-          id="day"
+          id="day-1"
           name="birthdate[day]"
           autocomplete="bday-day"
           aria-describedby="datefield_label_1"
@@ -45,7 +45,7 @@
           type="text"
           inputmode="numeric"
           pattern="[0-9]*"
-          id="year"
+          id="year-1"
           name="birthdate[year]"
           autocomplete="bday-year"
           aria-describedby="datefield_label_1"
@@ -71,7 +71,7 @@
             type="text"
             inputmode="numeric"
             pattern="[0-9]*"
-            id="month"
+            id="month-2"
             name="birthdate[month]"
             autocomplete="bday-month"
             aria-describedby="datefield_label_2"
@@ -86,7 +86,7 @@
             type="text"
             inputmode="numeric"
             pattern="[0-9]*"
-            id="day"
+            id="day-2"
             name="birthdate[day]"
             autocomplete="bday-day"
             aria-describedby="datefield_label_2"
@@ -101,7 +101,7 @@
             type="text"
             inputmode="numeric"
             pattern="[0-9]*"
-            id="year"
+            id="year-2"
             name="birthdate[year]"
             autocomplete="bday-year"
             aria-describedby="datefield_label_2"
