+++ packages/design-system/src/components/DateField/DateInput.test.jsx
@@ -1,0 +1,142 @@
+jest.mock('lodash.uniqueid', () => (str) => `${str}snapshot`);
+/* eslint-disable import/first */
+import { mount, shallow } from 'enzyme';
+import DateInput from './DateInput';
+import React from 'react';
+import renderer from 'react-test-renderer';
+
+describe('DateInput', () => {
+  it('renders with all defaultProps', () => {
+    expect(renderer.create(<DateInput />)).toMatchSnapshot();
+  });
+
+  it('is inversed', () => {
+    expect(renderer.create(<DateInput inversed />)).toMatchSnapshot();
+  });
+
+  it('has invalid month', () => {
+    expect(renderer.create(<DateInput monthInvalid />)).toMatchSnapshot();
+  });
+
+  it('has invalid day', () => {
+    expect(renderer.create(<DateInput dayInvalid />)).toMatchSnapshot();
+  });
+
+  it('has invalid year', () => {
+    expect(renderer.create(<DateInput yearInvalid />)).toMatchSnapshot();
+  });
+
+  it('has custom yearMax and yearMin', () => {
+    expect(renderer.create(<DateInput yearMax={2000} yearMin="1990" />)).toMatchSnapshot();
+  });
+
+  it('is disabled', () => {
+    expect(renderer.create(<DateInput disabled />)).toMatchSnapshot();
+  });
+
+  it('returns reference to input fields', () => {
+    const refs = {};
+    const props = {
+      dayDefaultValue: '1',
+      dayFieldRef: (el) => {
+        refs.day = el;
+      },
+      monthDefaultValue: '22',
+      monthFieldRef: (el) => {
+        refs.month = el;
+      },
+      yearDefaultValue: '3333',
+      yearFieldRef: (el) => {
+        refs.year = el;
+      },
+    };
+    mount(<DateInput {...props} />);
+
+    expect(refs.day.value).toBe(props.dayDefaultValue);
+    expect(refs.month.value).toBe(props.monthDefaultValue);
+    expect(refs.year.value).toBe(props.yearDefaultValue);
+  });
+
+  describe('event handlers', () => {
+    let props;
+
+    beforeEach(() => {
+      props = {
+        onBlur: jest.fn(),
+        onChange: jest.fn(),
+      };
+    });
+
+    it('does not require event handler', () => {
+      const wrapper = shallow(<DateInput />);
+
+      // This shouldn't result in an error
+      wrapper.find('TextField').at(0).simulate('blur');
+    });
+
+    it('calls onBlur when month is blurred', () => {
+      const wrapper = shallow(<DateInput {...props} />);
+
+      wrapper.find('TextField').at(0).simulate('blur');
+
+      expect(props.onBlur).toHaveBeenCalledTimes(1);
+      expect(props.onChange).not.toHaveBeenCalledTimes(1);
+    });
+
+    it('calls onChange when day is changed', () => {
+      const wrapper = shallow(<DateInput {...props} />);
+
+      wrapper.find('TextField').at(1).simulate('change');
+
+      expect(props.onBlur).not.toHaveBeenCalledTimes(1);
+      expect(props.onChange).toHaveBeenCalledTimes(1);
+    });
+
+    it('calls onComponentBlur when component loses focus', (done) => {
+      const onComponentBlur = jest.fn();
+      const wrapper = shallow(<DateInput onComponentBlur={onComponentBlur} />);
+
+      wrapper.find('TextField').at(2).simulate('blur');
+
+      setTimeout(() => {
+        expect(onComponentBlur).toHaveBeenCalledTimes(1);
+        done();
+      }, 30);
+    });
+
+    it('does not call onComponentBlur when focus switches to other date component', (done) => {
+      const onComponentBlur = jest.fn();
+      const wrapper = mount(<DateInput onComponentBlur={onComponentBlur} />);
+
+      wrapper.find('TextField').at(0).simulate('blur');
+
+      setTimeout(() => {
+        expect(onComponentBlur).not.toHaveBeenCalledTimes(1);
+        done();
+      }, 30);
+    });
+
+    it('formats the date as a single string', () => {
+      props = Object.assign(
+        {
+          dateFormatter: (values) => {
+            return `${values.month} ${values.day} ${values.year}`;
+          },
+          monthValue: '1',
+          dayValue: '22',
+          yearValue: '3333',
+        },
+        props
+      );
+
+      const wrapper = mount(<DateInput {...props} />);
+      wrapper.find('input').at(1).simulate('change');
+      wrapper.find('input').at(1).simulate('blur');
+
+      expect(props.onBlur).toHaveBeenCalledTimes(1);
+      expect(props.onChange).toHaveBeenCalledTimes(1);
+      expect(props.onBlur.mock.calls[0][1]).toBe('1 22 3333');
+      expect(props.onChange.mock.calls[0][1]).toBe('1 22 3333');
+    });
+  });
+});
